Clean up uss_qualifier and prepare to make SCD resources
