--- monitoring/uss_qualifier/rid/.gitignore
@@ -1,1 +1,0 @@
-test_definitions
