+++ monitoring/uss_qualifier/scd/executor/test_steps/delete_flight.py
@@ -1,5 +1,5 @@
 from monitoring.monitorlib.clients.scd_automated_testing import QueryError
-from monitoring.uss_qualifier.scd.data_interfaces import TestStep
+from monitoring.uss_qualifier.resources.flight_planning.automated_test import TestStep
 from monitoring.uss_qualifier.scd.executor.errors import TestRunnerError
 from monitoring.uss_qualifier.scd.executor.target import TestTarget
 from monitoring.uss_qualifier.scd.reports import TestStepReference
