+++ monitoring/uss_qualifier/scd/executor/test_steps/inject_flight.py
@@ -1,6 +1,6 @@
 from time import timezone
 
-from datetime import date, datetime, timedelta, timezone, tzinfo
+from datetime import datetime, timedelta, timezone
 from monitoring.monitorlib import scd
 from monitoring.monitorlib.clients.scd_automated_testing import QueryError
 from monitoring.monitorlib.fetch import scd as fetch_scd
@@ -14,7 +14,7 @@ from monitoring.uss_qualifier.common_data_definitions import (
     SubjectType,
     IssueSubject,
 )
-from monitoring.uss_qualifier.scd.data_interfaces import TestStep
+from monitoring.uss_qualifier.resources.flight_planning.automated_test import TestStep
 from monitoring.uss_qualifier.scd.executor.errors import TestRunnerError
 from monitoring.uss_qualifier.scd.executor.target import TestTarget
 from monitoring.uss_qualifier.scd.reports import Issue, TestStepReference
