--- monitoring/uss_qualifier/scd/simulator/README.md
@@ -1,15 +1,0 @@
-# USS Qualifier - SCD simulator
-
-This module implements the simulator which generates test definitions.
-
-## Run
-This module can be run using the following script:
-```
-../../bin/generate_scd_test_definition.sh <LOCALE>
-```
-
-## Test
-From this folder, run:
-```
-PYTHONPATH=$(pwd)/../../../.. pytest
-```
