+++ monitoring/uss_qualifier/test_data/che/flight_planning/flight-authorisation-validation-1.json
