+++ monitoring/uss_qualifier/test_data/che/flight_planning/nominal-planning-1.json
