+++ monitoring/uss_qualifier/test_data/che/flight_planning/nominal-planning-priority-1.json
